% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_event_date_possibilities_table.R
\name{build_event_date_possibilities_table}
\alias{build_event_date_possibilities_table}
\title{Build table of event date possibilities}
\usage{
build_event_date_possibilities_table(
  participant_level_data,
  omega.hat = build_omega_table(participant_level_data, bin_width = bin_width),
  ...
)
}
\arguments{
\item{participant_level_data}{a \link{data.frame}}
}
\value{
a data.frame
}
\description{
Build table of event date possibilities
}
