% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_phi_function_from_coefs}
\alias{build_phi_function_from_coefs}
\title{convert a pair of simple logistic regression coefficients into P(Y|T) curve:}
\usage{
build_phi_function_from_coefs(coefs)
}
\arguments{
\item{coefs}{numeric vector of coefficients}
}
\value{
function(t) P(Y=1|T=t)
}
\description{
convert a pair of simple logistic regression coefficients into P(Y|T) curve:
}
