% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_mu}
\alias{compute_mu}
\title{compute mean window period duration from simple logistic regression coefficients}
\usage{
compute_mu(theta)
}
\arguments{
\item{theta}{numeric vector of coefficients}
}
\value{
numeric scalar: mean window period duration
}
\description{
compute mean window period duration from simple logistic regression coefficients
}
