% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_omega.R
\name{graph_omega}
\alias{graph_omega}
\title{Graph seroconversion hazard model}
\usage{
graph_omega(omega)
}
\arguments{
\item{omega}{a \link{data.frame} containing parameter values for
the seroconversion hazard model}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Graph seroconversion hazard model
}
\examples{
example_model <-
  system.file("extdata", "example_model.rds", package = "rwicc") |>
  readRDS()
omega_est_EM <- example_model$Omega
omega_est_EM |> graph_omega()

}
