% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_S.R
\name{graph_S}
\alias{graph_S}
\title{Plot the estimated seroconversion-date distribution for one participant}
\usage{
graph_S(subject_level_data_possibilities, id = 1)
}
\arguments{
\item{subject_level_data_possibilities}{a \link{data.frame} of per-subject
seroconversion-date possibilities with their estimated probabilities}

\item{id}{the participant ID to plot}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot the estimated seroconversion-date distribution for one participant
}
\keyword{internal}
