% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_simulated_densities.R
\name{graph_simulated_densities}
\alias{graph_simulated_densities}
\title{Plot simulated seroconversion density curves}
\usage{
graph_simulated_densities()
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plots example seroconversion-time densities for several intercept/slope
hazard combinations.
}
\keyword{internal}
