% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_simulated_hazards.R
\name{graph_simulated_hazards}
\alias{graph_simulated_hazards}
\title{Plot simulated seroconversion hazard curves}
\usage{
graph_simulated_hazards()
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plots example linear seroconversion hazards for several intercept/slope
combinations.
}
\keyword{internal}
