% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_simulated_survival_curves.R
\name{graph_simulated_survival_curves}
\alias{graph_simulated_survival_curves}
\title{Plot simulated seroconversion survival curves}
\usage{
graph_simulated_survival_curves()
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plots example seroconversion-time survival curves for several
intercept/slope hazard combinations.
}
\keyword{internal}
