% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phi_curves.R
\name{plot_phi_curves}
\alias{plot_phi_curves}
\title{Plot true and estimated curves for P(Y=1|T=t)}
\usage{
plot_phi_curves(
  theta_true,
  theta.hat_joint,
  theta.hat_midpoint,
  theta.hat_uniform
)
}
\arguments{
\item{theta_true}{the coefficients of the data-generating model P(Y=1|T=t)}

\item{theta.hat_joint}{the estimated coefficients from the joint model}

\item{theta.hat_midpoint}{the estimated coefficients from midpoint imputation}

\item{theta.hat_uniform}{the estimated coefficients from uniform imputation}
}
\value{
a ggplot
}
\description{
Plot true and estimated curves for P(Y=1|T=t)
}
\examples{
vignette("how-to-use-rwicc")

}
