% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwicc-package.R
\docType{package}
\name{rwicc-package}
\alias{rwicc-package}
\title{rwicc: Regression with Interval-Censored Covariates}
\description{
Provides functions to simulate and analyze data for a regression model with an interval censored covariate, as described in Morrison et al. (2021) \doi{10.1111/biom.13472}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://d-morrison.github.io/rwicc/}
  \item \url{https://github.com/d-morrison/rwicc}
  \item Report bugs at \url{https://github.com/d-morrison/rwicc/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ron Brookmeyer
}

}
\keyword{internal}
