% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwicc.R
\docType{package}
\name{rwicc}
\alias{rwicc}
\title{rwicc: Regression with Interval-Censored Covariates}
\description{
The \code{rwicc} package implements a regression model with an
interval-censored covariate using an EM algorithm, as described in Morrison et al (2021); \doi{10.1111/biom.13472}.
}
\section{rwicc functions}{

The main \code{rwicc} functions are:
\itemize{
\item \code{\link{simulate_interval_censoring}}
\item \code{\link{fit_joint_model}}
}
}

\references{
Morrison, Laeyendecker, and Brookmeyer (2021).
"Regression with interval-censored covariates: Application to cross-sectional incidence estimation".
Biometrics. \doi{10.1111/biom.13472}.
}
