% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroconversion_survival_function.R
\name{seroconversion_density_function}
\alias{seroconversion_density_function}
\title{Seroconversion density function}
\usage{
seroconversion_density_function(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=seroconversion_survival_function]{seroconversion_survival_function()}} and
\code{\link[=seroconversion_hazard_function]{seroconversion_hazard_function()}}}
}
\value{
a numeric vector of density values
}
\description{
Probability density, \code{survival * hazard}.
}
\keyword{internal}
