% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroconversion_survival_function.R
\name{seroconversion_survival_function}
\alias{seroconversion_survival_function}
\title{Seroconversion survival function}
\usage{
seroconversion_survival_function(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[=seroconversion_cumhaz_function]{seroconversion_cumhaz_function()}}}
}
\value{
a numeric vector of survival probabilities
}
\description{
Survival probability, \verb{exp(-cumulative hazard)}.
}
\keyword{internal}
