% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_event_date_possibilities_table.R
\name{build_event_date_possibilities_table}
\alias{build_event_date_possibilities_table}
\title{Build table of event date possibilities}
\usage{
build_event_date_possibilities_table(
  participant_level_data,
  bin_width = 1,
  omega_hat = build_omega_table(participant_level_data, bin_width = bin_width)
)
}
\arguments{
\item{participant_level_data}{a \link{data.frame} with columns:
\itemize{
\item ID: participant identifier
\item Stratum: indicator for which population stratum the participant belongs to
\item E: study entry date
\item L: left censoring interval endpoint
\item R: right censoring interval endpoint
}}

\item{bin_width}{number of days per bin}

\item{omega_hat}{a \link{data.frame} from \code{\link[=build_omega_table]{build_omega_table()}} representing the
seroconversion hazard model.}
}
\value{
a \link{data.frame} with columns:
\itemize{
\item ID: participant identifier
\item Stratum: indicator for which population stratum the participant belongs to
\item S: possible seroconversion dates
}
}
\description{
Build table of event date possibilities
}
\examples{
library(dplyr)
simulate_interval_censoring()$pt_data |>
mutate(Stratum = 1) |>
build_event_date_possibilities_table()
}
