% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_omega_table.R
\name{build_omega_table}
\alias{build_omega_table}
\title{Build table of dates with possible seroconversions in dataset}
\usage{
build_omega_table(participant_level_data, bin_width = 1)
}
\arguments{
\item{participant_level_data}{a \link{data.frame}}

\item{bin_width}{number of days per bin}
}
\value{
a \link{data.frame}
}
\description{
Build table of dates with possible seroconversions in dataset
}
