% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_omega.R
\name{graph_omega}
\alias{graph_omega}
\title{Graph seroconversion hazard model}
\usage{
graph_omega(omega)
}
\arguments{
\item{omega}{a \link{data.frame} containing parameter values for
the seroconversion hazard model}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2}
}
\description{
Graph seroconversion hazard model
}
\examples{
example_model <-
  fs::path_package("rwicc", "extdata/example_model.rds")
EM_algorithm_outputs <- readRDS(example_model)
omega_est_EM <- EM_algorithm_outputs$Omega
omega_est_EM |> graph_omega()


}
