% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_censoring_data.R
\name{plot_censoring_data}
\alias{plot_censoring_data}
\title{Pl}
\usage{
plot_censoring_data(
  dataset = simulate_interval_censoring(),
  label.size = 5,
  point_size = 5,
  min_n_MAA = 5,
  use_shape = FALSE,
  s_vjust = 2,
  labelled_IDs = 2
)
}
\arguments{
\item{s_vjust}{test2}

\item{labelled_IDs}{test}
}
\value{
a ggplot
}
\description{
Pl
}
