% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_censoring_data.R
\name{plot_censoring_data}
\alias{plot_censoring_data}
\title{Title}
\usage{
plot_censoring_data(dataset = simulate_interval_censoring()$pt_data)
}
\arguments{
\item{dataset}{}
}
\description{
Title
}
