% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_censoring_data.R
\name{plot_censoring_data}
\alias{plot_censoring_data}
\title{Plot censoring data}
\usage{
plot_censoring_data(
  dataset,
  included_IDs = unique(dataset$pt_data$ID),
  label_size = 5,
  point_size = 5,
  s_vjust = 2,
  labelled_IDs = included_IDs,
  xmin = min(dataset$pt_data$E) - months(1),
  xmax = max(dataset$obs_data$O)
)
}
\arguments{
\item{dataset}{output from \code{\link[=simulate_interval_censoring]{simulate_interval_censoring()}}}

\item{included_IDs}{\link{character} \link{vector} of IDs from \code{dataset} to include}

\item{label_size}{\link{numeric}:
passed to \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}'s \code{size} argument}

\item{point_size}{\link{numeric}: passed to \link[ggplot2:geom_point]{ggplot2::geom_point}'s \code{size} argument}

\item{s_vjust}{passed to \link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel}'s \code{vjust} argument}

\item{labelled_IDs}{\link{character} \link{vector}}

\item{xmin}{minimum displayed value for x-axis}

\item{xmax}{maximum displayed value for x-axis
indicating which IDs to label events for}
}
\value{
a ggplot
}
\description{
Plot censoring data
}
