% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwicc-package.R, R/rwicc.R
\docType{package}
\name{rwicc-package}
\alias{rwicc-package}
\title{rwicc: Regression with Interval-Censored Covariates}
\description{
Provides functions to simulate and analyze data for a regression model with an interval censored covariate, as described in Morrison et al. (2021) \doi{10.1111/biom.13472}.

The \code{rwicc} package implements a regression model with an
interval-censored covariate using an EM algorithm, as described in Morrison et al (2021); \doi{10.1111/biom.13472}.
}
\section{rwicc functions}{

The main \code{rwicc} functions are:
\itemize{
\item \code{\link{simulate_interval_censoring}}
\item \code{\link{fit_joint_model}}
}
}

\references{
Morrison, Laeyendecker, and Brookmeyer (2021).
"Regression with interval-censored covariates: Application to cross-sectional incidence estimation".
Biometrics. \doi{10.1111/biom.13472}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://d-morrison.github.io/rwicc/}
  \item \url{https://github.com/d-morrison/rwicc}
  \item Report bugs at \url{https://github.com/d-morrison/rwicc/issues}
}


Useful links:
\itemize{
  \item \url{https://d-morrison.github.io/rwicc/}
  \item \url{https://github.com/d-morrison/rwicc}
  \item Report bugs at \url{https://github.com/d-morrison/rwicc/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ron Brookmeyer
}

}
\keyword{internal}
